Int 730 update tenant config end points to be consistent (#119)
* Updates to support new tenant config

* Update tenant controller

* Updated for new name field in tenant

* Update to support patientMRNTypeText

* Lint update

* Update based on PR comment

* Removed decoding on tenantMnemonics.
Updated EHR and Tenant controllers to use new DAO updates.
Updated liquibase image in docker-compose.

* Code cov updates

+++ src/main/kotlin/com/projectronin/interop/proxy/server/tenant/model/ProviderPool.kt
@@ -7,7 +7,7 @@ package com.projectronin.interop.proxy.server.tenant.model
  * @property poolId The ID of the pool the provider is associated with
  */
 data class ProviderPool(
-    val providerPoolId: Long,
+    val providerPoolId: Int,
     val providerId: String,
     val poolId: String
 )
